Anmols Backend LK,TL,IVR/dairaization_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

torch==2.1.2
torchaudio==2.1.2

pyannote.audio==3.1.1
speechbrain==1.0.0

huggingface_hub==0.23.4
numpy==1.26.4
matplotlib==3.8.4
